requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
APScheduler==3.11.0
requests==2.32.3
beautifulsoup4==4.13.3
lxml==5.3.1
tenacity==9.0.0
selenium==4.29.0
undetected-chromedriver==3.5.5
